+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # master
 
+# 0.1.11
+
+* Enable `ignoredModules` to be a function that returns an array. This allows the
+  list of ignored modules to be dynamic.
+
+# 0.1.10
+
+* Fix error reporting
+
+# 0.1.9
+
+* Improve error reporting for import errors
+
 # 0.1.8
 
 * Update dependency to follow symlinks correctly